src/mcp-server/requirements.txt
azure-functions
azure-cosmos
azure-identity
openai
